// src/types/database.types.ts
export type Database = {
  public: {
    Tables: {
      categories: {
        Row: {
          id: string
          is_deleted: boolean
          name: string
          parent_id: string | null
          shop_id: string
        }
        Insert: {
          id?: string
          is_deleted?: boolean
          name: string
          parent_id?: string | null
          shop_id: string
        }
        Update: {
          id?: string
          is_deleted?: boolean
          name?: string
          parent_id?: string | null
          shop_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "categories_parent_id_fkey"
            columns: ["parent_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "categories_shop_id_fkey"
            columns: ["shop_id"]
            isOneToOne: false
            referencedRelation: "shops"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
    }
// ...
  }
}